requirements.txt
django>=3.2
djangorestframework==3.15.2
pydantic==2.9.1
